requirements.txt
annotated-types==0.5.0
appdirs==1.4.4
beautifulsoup4==4.12.2
bibtexparser==1.4.0
bidict==0.23.1
blinker==1.6.2
blis==0.7.10
catalogue==2.0.9
certifi==2023.7.22
cfgv==3.4.0
charset-normalizer==3.2.0
click==8.1.7
confection==0.1.3
cymem==2.0.8
distlib==0.3.7
en-core-web-lg @ https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.6.0/en_core_web_lg-3.6.0-py3-none-any.whl#sha256=9012584a8fcf5bc61bedb13a4b511fb81ffff9f770958ee06366671be10632e9
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.6.0/en_core_web_sm-3.6.0-py3-none-any.whl#sha256=83276fc78a70045627144786b52e1f2728ad5e29e5e43916ec37ea9c26a11212
exceptiongroup==1.2.2
filelock==3.12.4
Flask==2.3.3
Flask-SocketIO==5.5.1
greenlet==2.0.2
h11==0.14.0
huggingface-hub==0.16.4
identify==2.5.29
idna==3.4
importlib-metadata==6.8.0
iniconfig==2.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
langcodes==3.3.0
lxml==4.9.3
MarkupSafe==2.1.3
mpmath==1.3.0
msgpack==1.0.5
murmurhash==1.0.10
networkx==3.1
nltk==3.8.1
nodeenv==1.8.0
numpy==1.26.0
pandas==2.2.3
pathy==0.10.2
Pillow==10.0.1
platformdirs==3.10.0
pluggy==1.5.0
pre-commit==3.4.0
preshed==3.0.9
protobuf==6.30.1
pydantic==2.3.0
pydantic_core==2.6.3
pynvim==0.4.3
pyparsing==3.1.1
PyQt6-Qt6==6.8.2
PyQt6-WebEngine
pytest==8.3.5
python-dateutil==2.9.0.post0
python-engineio==4.11.2
python-socketio==5.12.1
pytz==2025.2
PyYAML==6.0.1
regex==2023.8.8
requests==2.31.0
safetensors==0.3.3
scikit-learn==1.3.0
scipy==1.11.2
sentence-transformers==2.2.2
sentencepiece==0.1.99
simple-websocket==1.1.0
six==1.17.0
sklearn==0.0.post9
smart-open==6.4.0
soupsieve==2.5
spacy==3.6.1
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.7
sympy==1.12
thinc==8.1.12
threadpoolctl==3.2.0
tokenizers==0.13.3
tomli==2.2.1
torch==2.0.1
torchvision==0.15.2
tqdm==4.66.1
transformers==4.33.2
typer==0.9.0
types-requests==2.31.0.2
types-urllib3==1.26.25.14
typing_extensions==4.8.0
tzdata==2025.2
urllib3==2.0.4
virtualenv==20.24.5
wasabi==1.1.2
Werkzeug==2.3.7
wsproto==1.2.0
zipp==3.16.2
datasets==2.14.5
evaluate==0.4.0
